feat: supplier enhancement + fix 4 bugs from testing
- Supplier Enhancement: new models (supplier_category, supplier_contact, supplier_bank_account), controllers, routes, migrations, validation schemas, frontend tab layout with 4 tabs (Umum/Kontak/Pembelian/Pajak)
- Bug #4 (BLOCKER): Fix storeArray() returning [0] for empty stores; fix parseStoreField() type mismatch; add missing fields to createCategorySchema
- Bug #2: Remove category-based filtering in ingredient getProductNames endpoint that prevented reuse
- Bug #3: PO quantity accepts decimals - fix FE regex/inputMode, change DB column INTEGER to DECIMAL(10,4)
- Bug #1: Fix region seeder to download before truncating; fix seeder idempotency check

+++ db/models/purchaseOrderItem.js
@@ -28,7 +28,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: true
       },
       quantity: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(10, 4),
         allowNull: false
       },
       unit: {
@@ -44,7 +44,7 @@ module.exports = (sequelize, DataTypes) => {
         defaultValue: 0
       },
       receivedQuantity: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(10, 4),
         defaultValue: 0
       },
       conversionToBase: {
